refactor(cli): extract manifest shared primitives
